+++ backend/src/main/java/codezap/template/repository/ThumbnailSnippetRepository.java
@@ -1,8 +1,8 @@
-package codezap.thumbnailsnippet.repository;
+package codezap.template.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import codezap.thumbnailsnippet.domain.ThumbnailSnippet;
+import codezap.template.domain.ThumbnailSnippet;
 
 public interface ThumbnailSnippetRepository extends JpaRepository<ThumbnailSnippet, Long> {
 }
